API

API

ItemGraphs.ItemGraph — Type.

ItemGraph{T}([graph]; lazy=false)

Create a new ItemGraph with items of type T based on a graph from LightGraphs. graph can be omitted and a DiGraph will be used by default. If lazy is set to true, the paths between items will be computed on-demand when getpath is called. Otherwise all paths recomputed whenever a new edge is inserted.

source

ItemGraphs.getpath — Method.

getpath(graph, from, to)

Get the path between the items from and to. Will throw an ItemGraphsException if either from or to are not a part of graph or if there is no path between them.

source

LightGraphs.add_edge! — Method.

add_edge!(graph, from, to)

Add an edge between from and to to graph.

source

LightGraphs.add_vertex! — Method.

add_vertex!(graph, item)

Add item to graph.

source

LightGraphs.has_path — Method.

has_path(graph, from, to)

Return true if there is a path between from and to in graph.

source

LightGraphs.has_vertex — Method.

has_vertex(graph, item)

Return true if item is contained in graph.

source